src/shinet/requirements.txt
Django==4.0
environs~=8.0.0
Pillow==9.4.0
asgiref
psycopg2-binary==2.9.5
gunicorn==20.1.0
djangorestframework==3.14.0
drf-yasg==1.21.5
django-cors-headers==3.14.0
requests==2.28.2
beautifulsoup4==4.12.1
channels==3.0.5
